test: 금액입력 테스트 분리
